Make cells content as parameter
