+++ app/src/main/java/com/jp_funda/boxiful/views/components/calendar_heat_map/CalendarHeatmap.kt
@@ -1,22 +1,32 @@
 package com.jp_funda.boxiful.views.components.calendar_heat_map
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.Divider
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.DpSize
 import androidx.compose.ui.unit.dp
 import com.jp_funda.boxiful.utils.date.DateIterator
 import java.time.DayOfWeek
 import java.time.LocalDate
+import java.time.format.DateTimeFormatter
 import java.time.format.TextStyle
 import java.util.*
 
 @Composable
 fun CalendarHeatmap(
     cellLevelMap: Map<LocalDate, CalendarHeatmapLevel>,
+    cellPopupContentMap: Map<LocalDate, @Composable () -> Unit>? = null,
     startDate: LocalDate,
     endDate: LocalDate = LocalDate.now(),
     cellSize: DpSize,
@@ -84,22 +94,48 @@ fun CalendarHeatmap(
                             }
 
                             val cellLevel = cellLevelMap[date]
-                            if (cellLevel != null) {
-                                CalendarHeatmapCell(
-                                    date = date,
-                                    cellSize = cellSize,
-                                    cellPadding = cellPadding,
-                                    roundSize = roundSize,
-                                    level = cellLevel,
-                                )
-                            } else {
-                                CalendarHeatmapCell(
-                                    date = date,
-                                    cellSize = cellSize,
-                                    cellPadding = cellPadding,
-                                    roundSize = roundSize,
-                                    level = CalendarHeatmapLevel.Level0,
-                                )
+                            CalendarHeatmapCell(
+                                date = date,
+                                cellSize = cellSize,
+                                cellPadding = cellPadding,
+                                roundSize = roundSize,
+                                level = cellLevel,
+                            ) {
+                                Column(
+                                    modifier = Modifier
+                                        .clip(RoundedCornerShape(10.dp))
+                                        .widthIn(max = 200.dp)
+                                        .background(Color.DarkGray),
+                                    horizontalAlignment = Alignment.CenterHorizontally,
+                                ) {
+                                    val dateFormatter = DateTimeFormatter.ofPattern("yyyy年MM月dd日")
+                                    Spacer(modifier = Modifier.height(10.dp))
+                                    Text(
+                                        text = dateFormatter.format(date),
+                                        color = Color.LightGray,
+                                        style = MaterialTheme.typography.caption,
+                                        modifier = Modifier.padding(horizontal = 20.dp),
+                                    )
+                                    Divider(
+                                        modifier = Modifier
+                                            .fillMaxWidth()
+                                            .padding(5.dp)
+                                    )
+                                    Text(
+                                        text = "消費したカロリー 22kcal",
+                                        color = Color.LightGray,
+                                        style = MaterialTheme.typography.caption,
+                                        modifier = Modifier.padding(horizontal = 20.dp),
+                                    )
+                                    Spacer(modifier = Modifier.height(5.dp))
+                                    Text(
+                                        text = "Boxifulポイント 22ポイント",
+                                        color = Color.LightGray,
+                                        style = MaterialTheme.typography.caption,
+                                        modifier = Modifier.padding(horizontal = 20.dp),
+                                    )
+                                    Spacer(modifier = Modifier.height(10.dp))
+                                }
                             }
                         }
                     }
