+++ app/src/main/java/com/jp_funda/boxiful/views/components/calendar_heat_map/CalendarHeatmapCell.kt
@@ -3,15 +3,13 @@ package com.jp_funda.boxiful.views.components.calendar_heat_map
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Divider
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
@@ -22,7 +20,6 @@ import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import com.jp_funda.boxiful.views.components.PopupWindowDialog
 import java.time.LocalDate
-import java.time.format.DateTimeFormatter
 import kotlin.math.roundToInt
 
 @Composable
@@ -31,8 +28,9 @@ fun CalendarHeatmapCell(
     cellSize: DpSize,
     cellPadding: Dp,
     roundSize: Dp,
-    level: CalendarHeatmapLevel,
+    level: CalendarHeatmapLevel?,
     alpha: Float = 0.5f,
+    popupContent: @Composable () -> Unit = {},
 ) {
     val isShowDialog = remember { mutableStateOf(false) }
 
@@ -42,8 +40,8 @@ fun CalendarHeatmapCell(
                 .padding(cellPadding)
                 .size(cellSize)
                 .clip(RoundedCornerShape(roundSize))
-                .background(level.color.copy(alpha = minOf(1f, alpha + 0.3f)))
-                .border(width = 1.dp, color = level.color, shape = RoundedCornerShape(roundSize))
+                .background(level?.color?.copy(alpha = minOf(1f, alpha + 0.3f)) ?: Color.Transparent)
+                .border(width = 1.dp, color = level?.color ?: Color.Transparent, shape = RoundedCornerShape(roundSize))
                 .clickable { isShowDialog.value = true }
         ) {
             PopupWindowDialog(
@@ -53,40 +51,7 @@ fun CalendarHeatmapCell(
                     cellSize.height.toPx().roundToInt(),
                 )
             ) {
-                // TODO content
-                Column(
-                    modifier = Modifier
-                        .clip(RoundedCornerShape(10.dp))
-                        .widthIn(max = 200.dp)
-                        .background(Color.DarkGray),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                ) {
-                    val dateFormatter = DateTimeFormatter.ofPattern("yyyy年MM月dd日")
-                    Spacer(modifier = Modifier.height(10.dp))
-                    Text(
-                        text = dateFormatter.format(date),
-                        color = Color.LightGray,
-                        style = MaterialTheme.typography.caption,
-                        modifier = Modifier.padding(horizontal = 20.dp),
-                    )
-                    Divider(modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(5.dp))
-                    Text(
-                        text = "消費したカロリー 22kcal",
-                        color = Color.LightGray,
-                        style = MaterialTheme.typography.caption,
-                        modifier = Modifier.padding(horizontal = 20.dp),
-                    )
-                    Spacer(modifier = Modifier.height(5.dp))
-                    Text(
-                        text = "Boxifulポイント 22ポイント",
-                        color = Color.LightGray,
-                        style = MaterialTheme.typography.caption,
-                        modifier = Modifier.padding(horizontal = 20.dp),
-                    )
-                    Spacer(modifier = Modifier.height(10.dp))
-                }
+                popupContent()
             }
         }
     }
